refactor: add homePage folder
홈 페이지에서 사용할 컴포넌트가 많을 것 같아, 모아둘 폴더를 추가함
